chore: updated view-fees to be a separate page
